+++ lib/app/data/models/money_objects/securities/security.dart
@@ -1,6 +1,7 @@
 // ignore_for_file: unnecessary_this
 import 'package:money/app/core/helpers/date_helper.dart';
 import 'package:money/app/core/helpers/list_helper.dart';
+import 'package:money/app/data/models/money_objects/stock_splits/stock_split.dart';
 import 'package:money/app/data/storage/data/data.dart';
 import 'package:money/app/modules/home/sub_views/view_stocks/picker_security_type.dart';
 
@@ -190,6 +191,13 @@ class Security extends MoneyObject {
     },
   );
 
+  FieldInt splitsCount = FieldInt(
+    name: 'Splits',
+    useAsColumn: false,
+    getValueForDisplay: (final MoneyObject instance) => (instance as Security).splitsHistory.length,
+  );
+
+  List<StockSplit> splitsHistory = [];
   // 2
   FieldString symbol = FieldString(
     name: 'Symbol',
@@ -233,6 +241,7 @@ class Security extends MoneyObject {
         tmp.lastPrice,
         tmp.cuspid,
         tmp.securityType,
+        tmp.splitsCount,
         tmp.numberOfTrades,
         tmp.holdingShares,
         tmp.holdingValue,